config.py
from dotenv import load_dotenv
import os

load_dotenv()  # Загрузка переменных из .env файла

BOT_TOKEN = os.getenv('BOT_TOKEN')
YOOKASSA_SHOP_ID = os.getenv('YOOKASSA_SHOP_ID')
YOOKASSA_SECRET_KEY = os.getenv('YOOKASSA_SECRET_KEY')
WIREGUARD_BASE_URL = os.getenv('WIREGUARD_BASE_URL')
WIREGUARD_PASSWORD = os.getenv('WIREGUARD_PASSWORD')