pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "aiogram>=3.13.1",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.10.4",
    "cairosvg>=2.8.2",
    "greenlet>=3.1.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
    "yookassa>=3.5.0",
]
